+++ Navigator/Models/DisplayWrapper.cs
@@ -1,0 +1,12 @@
+﻿namespace Navigator.Models
+{
+    public class DisplayWrapper<T>
+    {
+        public T WrappedModel;
+
+        public DisplayWrapper(T wrappedModel)
+        {
+            WrappedModel = wrappedModel;
+        }
+    }
+}
