+++ Navigator/Models/LoyaltyRewardsViewModel.cs
@@ -1,30 +1,15 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using EVEStandard.Models;
-using Navigator.DAL.SDE;
 
 namespace Navigator.Models
 {
     public class LoyaltyRewardsViewModel 
     {
-        public List<LoyaltyStoreOfferDisplayWrapper> Offers;
+        public List<ItemDisplayWrapper> Offers;
 
-        public LoyaltyRewardsViewModel(List<LoyaltyStoreOfferDisplayWrapper> offers)
+        public LoyaltyRewardsViewModel(List<ItemDisplayWrapper> offers)
         {
             Offers = offers.OrderByDescending(x => x.WrappedModel.LpCost).ToList();
-        }
-    }
-
-    public class LoyaltyStoreOfferDisplayWrapper : DisplayWrapper<LoyaltyStoreOffer>
-    {
-        private InvTypes _itemData;
-
-        public string ItemName => _itemData.TypeName;
-
-        public LoyaltyStoreOfferDisplayWrapper(InvTypes itemData, LoyaltyStoreOffer offer)
-            : base(offer)
-        {
-            _itemData = itemData;
         }
     }
 }
