+++ Navigator/Models/LoyaltyViewModel.cs
@@ -1,7 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using EVEStandard.Models;
-using Navigator.DAL.SDE;
 
 namespace Navigator.Models
 {
@@ -14,31 +12,4 @@ namespace Navigator.Models
             LoyaltyPoints = loyaltyPoints;
         }
     }
-
-    public class LoyaltyPointDisplayWrapper : DisplayWrapper<LoyaltyPoints>
-    {
-        private InvUniqueNames _npcCorpData;
-
-        public int LoyaltyPointQuantity => WrappedModel.Points;
-
-        public string CorpName => _npcCorpData.ItemName;
-
-        public LoyaltyPointDisplayWrapper(InvUniqueNames npcCorpData, LoyaltyPoints lp) 
-            : base(lp)
-        {
-            _npcCorpData = npcCorpData;
-        }
-    }
-
-    public class DisplayWrapper<T>
-    {
-        public T WrappedModel;
-
-        public DisplayWrapper(T wrappedModel)
-        {
-            WrappedModel = wrappedModel;
-        }
-    }
-
-
 }
